requirements.txt
discord.py
pyyaml
pygithub
redis
booktype
colorama
requests